+++ cmd/file.go
@@ -10,8 +10,8 @@ import (
 	"time"
 
 	spinner "github.com/briandowns/spinner"
+	"github.com/hahwul/dalfox/v2/internal/printing"
 	model "github.com/hahwul/dalfox/v2/pkg/model"
-	"github.com/hahwul/dalfox/v2/pkg/printing"
 	"github.com/hahwul/dalfox/v2/pkg/scanning"
 	voltFile "github.com/hahwul/volt/file"
 	voltHar "github.com/hahwul/volt/format/har"
