+++ cmd/sxss.go
@@ -1,7 +1,7 @@
 package cmd
 
 import (
-	"github.com/hahwul/dalfox/v2/pkg/printing"
+	"github.com/hahwul/dalfox/v2/internal/printing"
 	"github.com/hahwul/dalfox/v2/pkg/scanning"
 	"github.com/spf13/cobra"
 )
