+++ internal/generating/bulk.go
