+++ internal/generating/bulk_test.go
