+++ internal/har/writer.go
