+++ internal/optimization/abstraction.go
