+++ internal/optimization/inspectionParam.go
