+++ internal/optimization/optimization.go
@@ -10,7 +10,7 @@ import (
 	"os"
 	"strings"
 
-	"github.com/hahwul/dalfox/v2/pkg/har"
+	"github.com/hahwul/dalfox/v2/internal/har"
 
 	"github.com/hahwul/dalfox/v2/pkg/model"
 )
