+++ internal/optimization/replace.go
