+++ internal/optimization/replace_test.go
