+++ internal/printing/banner_test.go
