+++ internal/printing/codeview.go
@@ -1,4 +1,4 @@
-package scanning
+package printing
 
 import (
 	"strconv"
