+++ internal/printing/logger_test.go
