+++ internal/printing/multispin_test.go
