+++ internal/printing/util.go
