+++ internal/verification/verify.go
