+++ lib/interface.go
@@ -3,7 +3,7 @@ package lib
 import (
 	"time"
 
-	"github.com/hahwul/dalfox/v2/pkg/har"
+	"github.com/hahwul/dalfox/v2/internal/har"
 	"github.com/hahwul/dalfox/v2/pkg/model"
 )
