+++ pkg/scanning/staticAnlaysis.go
@@ -5,8 +5,8 @@ import (
 	"strconv"
 	"strings"
 
+	"github.com/hahwul/dalfox/v2/internal/optimization"
 	"github.com/hahwul/dalfox/v2/pkg/model"
-	"github.com/hahwul/dalfox/v2/pkg/optimization"
 )
 
 // StaticAnalysis is found information on original req/res
